# update API in preparation for format changes