--- a/export_test.go
+++ b/export_test.go
@@ -1,10 +1,5 @@
 package macaroon
 
-// Data returns the macaroon's data.
-func (m *Macaroon) Data() []byte {
-	return m.data
-}
-
 // AddThirdPartyCaveatWithRand adds a third-party caveat to the macaroon, using
 // the given source of randomness for encrypting the caveat id.
 var AddThirdPartyCaveatWithRand = (*Macaroon).addThirdPartyCaveatWithRand